Required stories

+++ app.py
@@ -3,26 +3,41 @@ from flask import Flask, render_template, request, redirect, url_for, flash
 from flask_login import LoginManager, UserMixin, login_required, login_user, logout_user, current_user
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_sqlalchemy import SQLAlchemy
-
+from flask_migrate import Migrate
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, ValidationError, TextField, TextAreaField
 from wtforms.validators import DataRequired, Email, EqualTo, Length, InputRequired
-
+from wtforms.widgets import TextArea
 app = Flask(__name__)
 app.config['SEND_FILE_MAX_AGE_DEFAULT'] = 0
 app.config['SECRET_KEY'] = "secret"
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+
+
+POSTGRES = {
+    'user': 'quyen',
+    'pw': '123',
+    'db': 'blog',
+    'host': 'localhost',
+    'port': 5432,
+}
+app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://%(user)s:%(pw)s@%(host)s:\
+%(port)s/%(db)s' % POSTGRES
+
 login = LoginManager(app)
 login.login_view = "login"
 db = SQLAlchemy(app)
 
+migrate = Migrate(app, db)
+
 
 class User(UserMixin, db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    user_name = db.Column(db.String(), nullable=False)
+    user_name = db.Column(db.String, nullable=False)
     email = db.Column(db.String(128), index=True, unique=True)
     password_hash = db.Column(db.String(128), nullable=False)
     posts = db.relationship('Post', backref='author', lazy='dynamic')
+    comments = db.relationship('Comment', backref='author', lazy='dynamic')
 
     def set_password(self, password):
         self.password_hash = generate_password_hash(password)
@@ -39,18 +54,59 @@ class Post(db.Model):
     created_on = db.Column(db.DateTime, server_default=db.func.now())
     updated_on = db.Column(
         db.DateTime, server_default=db.func.now(), server_onupdate=db.func.now())
-    user = db.relationship('User', backref='post')
+    comments = db.relationship("Comment", backref="post", lazy="dynamic")
+
+
+class Comment(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    body = db.Column(db.String(255), nullable=False)
+    created_on = db.Column(db.DateTime, server_default=db.func.now())
+    author_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    post_id = db.Column(db.Integer, db.ForeignKey('post.id'), nullable=False)
+
+
+db.create_all()
 
 
 class NewPost(FlaskForm):
     title = StringField("Article Title", validators=[
-                        DataRequired(), Length(min=5, max=255)])
+                        DataRequired(), Length(min=5, max=50, message="Title must be 5-50 characters long")])
     body = TextField("Article Body", validators=[
-        DataRequired(), Length(min=5)])
+        DataRequired(), Length(min=5, message="Body must be at least 5 characters")], widget=TextArea())
     submit = SubmitField("Submit")
 
 
-db.create_all()
+class Login(FlaskForm):
+    email = StringField("Email", validators=[Email(
+        message="Please enter a valid email!"), DataRequired()])
+    password = PasswordField("Password", validators=[DataRequired()])
+    submit = SubmitField("Login")
+
+
+class SignUp(FlaskForm):
+    email = StringField("Email", validators=[
+                        DataRequired(), Email(message='Please enter a valid email')])
+    username = StringField("User Name", validators=[DataRequired(), Length(
+        max=25, min=3, message='User name must be 3-23 characters long')])
+    password = PasswordField("Password", validators=[DataRequired(), Length(
+        min=3, message="Password must be at least 3 characters")])
+    cf_password = PasswordField(
+        "Re-enter Password", validators=[DataRequired(), EqualTo("password", message="Password not matching")])
+    submit = SubmitField("Sign Up")
+
+    def validate_email(self, field):
+        if User.query.filter_by(email=field.data).first():
+            raise ValidationError("Your email has been registered")
+
+    def validate_username(self, field):
+        if User.query.filter_by(user_name=field.data).first():
+            raise ValidationError("Your username has been registered")
+
+
+class NewComment(FlaskForm):
+    body = StringField("Your comment here...", validators=[DataRequired(message="Data required"), Length(
+        max=255, message="Maximum 255 characters")], widget=TextArea())
+    submit = SubmitField("Comment")
 
 # User =======================================================
 @login.user_loader
@@ -63,25 +119,35 @@ def home():
     return render_template("layout.html")
 
 
+@app.errorhandler
+def errorhandler():
+    return "Error"
+
+
 @app.route("/login", methods=["POST", "GET"])
 def login():
+    form = Login()
     if request.method == "POST":
-        user = User.query.filter_by(email=request.form['email']).first()
-        if user is not None and user.check_password(request.form['password']):
-            flash("Logged in successfully!", 'success')
-            login_user(user)
-            return redirect(url_for("profile"))
-        flash('Sorry, your username or password is incorrect.', "danger")
+        if form.validate_on_submit():
+            user = User.query.filter_by(email=form.email.data).first()
+            if user and user.check_password(form.password.data):
+                login_user(user)
+                flash("Logged in successfully!", 'success')
+                return redirect(url_for("posts"))
+            flash("Email or password is not correct", "danger")
+        for field_name, errors in form.errors.items():
+            flash(errors[0])
         return redirect(url_for("login"))
     else:
-        return render_template("login.html")
+        return render_template("login.html", form=form)
 
 
-@app.route("/profile")
+@app.route("/profile/<id>")
 @login_required
-def profile():
-    posts_count = Post.query.filter_by(author_id=current_user.id).count()
-    return render_template("profile.html", user=current_user, count=posts_count)
+def profile(id):
+    user = User.query.filter_by(id=id).first()
+    posts_count = Post.query.filter_by(author_id=id).count()
+    return render_template("profile.html", user=user, count=posts_count)
 
 
 @app.route('/logout')
@@ -92,24 +158,33 @@ def logout():
 
 @app.route("/register", methods=["POST", "GET"])
 def sign_up():
+    form = SignUp()
     if request.method == "POST":
-        check_dup = User.query.filter_by(email=request.form['email']).first()
-        if check_dup is None:
-            new_user = User(
-                email=request.form['email'], user_name=request.form['name'])
-            new_user.set_password(request.form['password'])
+        if form.validate_on_submit():
+            new_user = User(user_name=form.username.data,
+                            email=form.email.data)
+            new_user.set_password(form.password.data)
             db.session.add(new_user)
             db.session.commit()
             return redirect(url_for("login"))
         else:
-            flash("Sign up not successful! Email already exists", "danger")
-    return render_template("signup.html")
+            for field_name, errors in form.errors.items():
+                flash(errors[0])
+    return render_template("signup.html", form=form)
 
 # Posts ================================================================
 @app.route("/posts")
-@login_required
 def posts():
-    posts = Post.query.filter_by(author_id=current_user.id).all()
+    # import code
+    # code.interact(local=dict(globals(), **locals()))
+
+    posts = Post.query.all()
+    return render_template("posts.html", posts=posts)
+
+
+@app.route("/user/<id>/posts")
+def user_posts(id):
+    posts = Post.query.filter_by(author_id=id).all()
     return render_template("posts.html", posts=posts)
 
 
@@ -120,38 +195,98 @@ def upload():
     if request.method == "POST":
         if form.validate_on_submit():
             new_post = Post(title=form.title.data,
-                            body=form.body.data, author_id=current_user.id)
+                            body=form.body.data)
+            current_user.posts.append(new_post)
             db.session.add(new_post)
             db.session.commit()
             flash("Post successfully uploaded!", 'success')
             return redirect(url_for("posts"))
-    return render_template("create_form.html", form=form)
+    return render_template("new_post.html", form=form)
+
+
+@app.route('/post/<id>', methods=['POST', "GET"])
+def post(id):
+    post = Post.query.filter_by(id=id).first()
+    form = NewComment()
+    if request.method == "POST":
+        if current_user.is_authenticated:
+            if form.validate_on_submit():
+                new_comment = Comment(body=form.body.data,
+                                      created_on=datetime.now())
+                current_user.comments.append(new_comment)
+                post.comments.append(new_comment)
+                db.session.add(new_comment)
+                db.session.commit()
+        else:
+            flash("You must sign in to comment")
+            for field_name, errors in form.errors.items():
+                flash(errors[0])
+        return redirect(url_for('post', id=id))
+    return render_template('single_post.html', post=post, comment_form=form, comments=post.comments.all())
 
 
 @app.route("/edit/<id>", methods=["POST", "GET"])
 @login_required
 def edit(id):
     form = NewPost()  # Reuse form for post creation
-    post = Post.query.filter_by(id=id).first()
+    post = Post.query.filter_by(id=id, author_id=current_user.id).first()
     if request.method == "POST":
-        if form.validate_on_submit():
+        if form.validate_on_submit() and post:
             post.title = form.title.data
             post.body = form.body.data
             post.updated_on = datetime.now().now()
             db.session.commit()
             flash("Post successfully edited!", 'success')
-            return redirect(url_for("posts"))
+        else:
+            flash("You are not authorized to edit this post", "danger")
+            for field_name, errors in form.errors.items():
+                flash(errors[0])
+        return redirect(url_for("posts"))
     return render_template("edit_form.html", form=form, post=post)
 
 
-@app.route("/delete/<id>", methods=["POST", "GET"])
+@app.route("/edit-comment/<id>", methods=['POST', "GET"])
+@login_required
+def edit_comment(id):
+    form = NewComment()
+    comment = Comment.query.filter_by(id=id).first()
+    if current_user.id == comment.author_id:
+        if request.method == "POST" and form.validate_on_submit():
+            comment.body = form.body.data
+            db.session.commit()
+            return redirect(url_for("post", id=comment.post_id))
+        for field_name, errors in form.errors.items():
+            flash(errors[0])
+        return render_template("edit_comment.html", comment=comment, form=form)
+    else:
+        flash("You are not authorized to edit this comment", "danger")
+    return redirect(url_for("post", id=comment.post_id))
+
+
+@app.route("/delete/<id>", methods=["GET"])
 @login_required
 def delete(id):
-    post = Post.query.filter_by(id=id).first()
-    db.session.delete(post)
-    db.session.commit()
-    flash("Post successfully deleted!", 'success')
+    post = Post.query.filter_by(id=id, author_id=current_user.id).first()
+    if post:
+        db.session.delete(post)
+        db.session.commit()
+        flash("Post successfully deleted!", 'success')
+    else:
+        flash("You are not authorized to delete this post", "danger")
     return redirect(url_for("posts"))
+
+
+@app.route("/delete-comment/<id>")
+@login_required
+def delete_comment(id):
+    comment = Comment.query.filter_by(id=id, author_id=current_user.id).first()
+    if comment:
+        db.session.delete(comment)
+        db.session.commit()
+        flash("Comment deleted!")
+    else:
+        flash("You are not authorized to delete this comment", "danger")
+    return redirect(url_for("post", id=comment.post_id))
 
 
 if __name__ == '__main__':
